refactor: [FE] 파일 위치 변경
- 하나의 컴포넌트로 사용하기 위해 파일 위치 변경
- Tabmenu -> ContentEditor
- ContentEditor -> Writer
