+++ frontend/src/components/index.js
@@ -1,4 +1,3 @@
 export { default as LoginForm } from "./LoginForm/LoginForm";
 export { default as Button } from "./Button/Button";
 export { default as ContentEditor } from "./ContentEditor/ContentEditor";
-export { default as TabMenu } from "./TabMenu/TabMenu";
